Add logging and framework object injection
When we export the framework for those that require it, we now inject instances
of the exported framework object into objects that are instaniated from the
exported classes. This allows all internal classes to obtain an instance of the
framework itself so that they can use user-configured objects like loggers,
etc. This is a poor-man's way of allowing dependency injection.